updated init screen, modularized post
