+++ src/actions/actions-type.js
@@ -1,3 +1,7 @@
 export const FETCH_POST = "FETCH_POST"
 export const UPDATE_LIKES_POST = "UPDATE_LIKES_POST"
 export const UPDATE_USER_DETAILS = "UPDATE_USER_DETAILS"
+
+export const GET_POST_REQUEST = "GET_POST_REQUEST"
+export const GET_POST_SUCCESS = "GET_POST_SUCCESS"
+export const GET_POST_ERROR = "GET_POST_ERROR"
