+++ src/globals.js
@@ -1,2 +1,3 @@
 
 export const tintColor = '#0960BD'
+export const BASE_URL = 'https://backtick-api.herokuapp.com'
