+++ src/route.js
@@ -19,140 +19,140 @@ import VerifyCode from './containers/auth/forgotpassword/askcodeandpassword'
 import ConfirmEmail from './containers/auth/forgotpassword/sendemail'
 import SignUp from './containers/auth/signup'
 import Auth from './containers/auth/'
+import { tintColor } from './globals';
 
 export const Home = TabNavigator(
-    {
-        CreatePost:{
-            screen:CreatePost
-        },
-        Bulletin:{
-            screen:Bulletin
-        },
-        Notification:{
-            screen:Notification
-        },
-        Schedules:{
-            screen:Bulletin
-        },
-        Saved:{
-            screen:Save
-        },
+{
+	CreatePost:{
+			screen:CreatePost
+	},
+	Bulletin:{
+			screen:Bulletin
+	},
+	Notification:{
+			screen:Notification
+	},
+	Schedules:{
+			screen:Bulletin
+	},
+	Saved:{
+			screen:Save
+	},
 
-    },
-    {
-        navigationOptions: ({ navigation }) => ({
-          tabBarIcon: ({ focused, tintColor }) => {
-            const { routeName } = navigation.state;
-            let iconName;
-            if (routeName === 'CreatePost')
-              iconName = `ios-add-circle${focused ? '' : '-outline'}`;
-            else if (routeName === 'Bulletin')
-              iconName = `ios-albums${focused ? '' : '-outline'}`;
-            else if(routeName === 'Notification')
-                iconName = `ios-notifications${focused ? '' : '-outline'}`;
-            else if(routeName === 'Schedules')
-                iconName = `ios-calendar${focused ? '' : '-outline'}`
-            else if(routeName === 'Saved')
-                iconName = `ios-bookmark${focused ? '' : '-outline'}`
-    
-            // You can return any component that you like here! We usually use an
-            // icon component from react-native-vector-icons
-            return <Icon name={iconName} style={{color: tintColor}}/>;
-          },
-          tabStyle :{
-          }
-        }),
-        tabBarOptions: {
-          activeTintColor: '#0960BD',
-          inactiveTintColor: '#666',
-          pressColor :'red',
-          showLabel: true,
-          style: {
-            backgroundColor: 'white',
-          }
+},
+{
+	navigationOptions: ({ navigation }) => ({
+		tabBarIcon: ({ focused, tintColor }) => {
+			const { routeName } = navigation.state;
+			let iconName;
+			if (routeName === 'CreatePost')
+				iconName = `ios-add-circle${focused ? '' : '-outline'}`;
+			else if (routeName === 'Bulletin')
+				iconName = `ios-albums${focused ? '' : '-outline'}`;
+			else if(routeName === 'Notification')
+					iconName = `ios-notifications${focused ? '' : '-outline'}`;
+			else if(routeName === 'Schedules')
+					iconName = `ios-calendar${focused ? '' : '-outline'}`
+			else if(routeName === 'Saved')
+					iconName = `ios-bookmark${focused ? '' : '-outline'}`
 
-        },
-        tabBarComponent: TabBarBottom,
-        initialRouteName :'Bulletin',
-        tabBarPosition: 'bottom',
-        animationEnabled: true,
-        swipeEnabled: true,
-        allowFontScaling:true
-      }
+			// You can return any component that you like here! We usually use an
+			// icon component from react-native-vector-icons
+			return <Icon name={iconName} style={{color: tintColor}}/>;
+		},
+		tabStyle :{
+		}
+	}),
+	tabBarOptions: {
+		activeTintColor: tintColor,
+		inactiveTintColor: '#666',
+		pressColor :'red',
+		showLabel: true,
+		style: {
+			backgroundColor: 'white',
+		}
+
+	},
+	tabBarComponent: TabBarBottom,
+	initialRouteName :'Bulletin',
+	tabBarPosition: 'bottom',
+	animationEnabled: true,
+	swipeEnabled: true,
+	allowFontScaling:true
+}
 )
 
 
 export const SingedIn = StackNavigator(
-    {   
+{   
 
-        Home:{
-            screen:Home
-        }
-    },
-    {
-        initialRouteName:'Home'
-    }
+	Home:{
+			screen:Home
+	}
+},
+{
+	initialRouteName:'Home'
+}
 )
 
 export const Join = StackNavigator({
-    Join:{
-        screen:JoinClass
-    }
+Join:{
+	screen:JoinClass
+}
 })
 export const SingedOut = StackNavigator(
-    {
-        Auth:{
-            screen:Auth
-        },
-        SignUp:{
-            screen:SignUp
-        },
-        VerifyCode:{
-            screen:VerifyCode
-        },
-        ConfirmEmail:{
-            screen:ConfirmEmail
-        }
-    },
-    {
-        initialRouteName:'Auth'
-    }
+{
+	Auth:{
+			screen:Auth
+	},
+	SignUp:{
+			screen:SignUp
+	},
+	VerifyCode:{
+			screen:VerifyCode
+	},
+	ConfirmEmail:{
+			screen:ConfirmEmail
+	}
+},
+{
+	initialRouteName:'Auth'
+}
 )
 
 export const InitScreen = StackNavigator(
-    {   
-        Init:{
-            screen:Init
-        },
-        UpdateInfo:{
-            screen:UpdateInfo
-        }
-    }
+{   
+	Init:{
+			screen:Init
+	},
+	UpdateInfo:{
+			screen:UpdateInfo
+	}
+}
 )
 export const RootNavigator =  SwitchNavigator(
-        {   
-            AuthLoading:{
-                screen:AuthLoading
-            },
-            SignedIn:{
-                screen : SingedIn
-            },
-            SignedOut:{
-                screen : SingedOut
-            },
-            VerifyEmail:{
-                screen:VerifyEmail
-            },
-            InitScreen:{
-                screen:InitScreen
-            },
-            Join:{
-                screen:Join
-            }
-    
-        },
-        {
-            initialRouteName:'AuthLoading'
-        }
-    )
- 
+	{   
+			AuthLoading:{
+					screen:AuthLoading
+			},
+			SignedIn:{
+					screen : SingedIn
+			},
+			SignedOut:{
+					screen : SingedOut
+			},
+			VerifyEmail:{
+					screen:VerifyEmail
+			},
+			InitScreen:{
+					screen:InitScreen
+			},
+			Join:{
+					screen:Join
+			}
+
+	},
+	{
+			initialRouteName:'AuthLoading'
+	}
+)
